% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{data}
\name{arguments_arima}
\alias{arguments_arima}
\title{Arguments for auto arima models}
\format{tibble}
\usage{
data(arguments_arima)
}
\description{
Tibble with the following columns
}
\details{
\itemize{
\item pars_id id of parameter combination
\item expand_window_by number of observations added to basic window length.
Usually this number is equal to the number of parameters in the model. Three for an average ARIMA model.
Can not be computed beforehand for auto.arima because the number of parameters is not known a-priori.
}
}
\keyword{datasets}
