% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{data}
\name{arguments_ets}
\alias{arguments_ets}
\title{Arguments for ETS models}
\format{tibble}
\usage{
data(arguments_ets)
}
\description{
Tibble with the following columns
}
\details{
\itemize{
\item pars_id id of parameter combination
\item expand_window_by number of observations added to basic window length.
Usually this number is equal to the number of parameters in the model. Three for an average ETS model.
Can not be computed beforehand for \code{ets} because the number of parameters is not known a-priori.
}
}
\keyword{datasets}
