% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{data}
\name{arguments_rw}
\alias{arguments_rw}
\title{Arguments for random walk models}
\format{tibble}
\usage{
data(arguments_rw)
}
\description{
Tibble with the following columns
}
\details{
\itemize{
\item pars_id id of parameter combination
\item expand_window_by number of observations added to basic window length.
Usually this number is equal to the number of parameters in the model. One for RW model with drift.
}
}
\keyword{datasets}
