% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{data}
\name{arguments_var_lasso_toy}
\alias{arguments_var_lasso_toy}
\title{Arguments for VAR-lasso models toy version}
\format{tibble}
\usage{
data(arguments_var_lasso_toy)
}
\description{
Tibble with the following columns
}
\details{
\itemize{
\item p number of lags
\item struct LASSO lag structure
Three top players according to \url{https://arxiv.org/pdf/1508.07497.pdf} page 23 are
SparseLag (Lag Sparse Group VARX-L), OwnOther (Own/Other Group VARX-L), SparseOO (Own/Other Sparse Group VARX-L).
but SparseLag and SparseOO hangs the computer.
\item gran_1 first granularity parameter: 
left grid search border is equal to maximal lambda / gran_1. 
Maximal labmda sets all coefficient estimates to zero.
\item gran_2 second granularity parameter: number of grid points for cross-validation.
Values gran_1 = 25 and grand_2 = 10 are suggested by BigVAR user's guide, 
\url{http://www.wbnicholson.com/BigVAR.pdf}, page 9.
\item pars_id id of parameter combination
\item expand_window_by number of observations added to basic window length.
Usually this number is equal to the number of parameters in the model. Three for an average ARIMA model.
Can not be computed beforehand for auto.arima because the number of parameters is not known a-priori.
}
See also full version, \code{arguments_var_lasso}.
}
\keyword{datasets}
