% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{attempt_to_status}
\alias{attempt_to_status}
\title{Extract message from forecasting attempt}
\usage{
attempt_to_status(attempt)
}
\arguments{
\item{attempt}{estimation attempt.}
}
\value{
character message
}
\description{
\code{attempt_to_status} extracts error message from estimation attempt.
}
\details{
Gives "OK" if attempt was successful.
}
\examples{
# TODO
}
