% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_evaluation.R
\name{augment_forecasts}
\alias{augment_forecasts}
\title{Attach real y and model information to all_forecasts data set.}
\usage{
augment_forecasts(all_forecasts, fits_long, y)
}
\arguments{
\item{all_forecasts}{tibble that contains row_number, model_id, variable and value columns.}

\item{fits_long}{tibble that contains T_end, model_type and model_filename columns}

\item{y}{actual time series}
}
\value{
all_forecasts with new columns: y_actual, var_set, pars_id...
}
\description{
Add new columns to all_forecasts with short model description and actual y.
}
\details{
Original all_forecasts contains row_number, model_id, variable and value columns.
Using info from fits_long this function attach actual y.
}
