% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_evaluation.R
\name{calculate_rmse}
\alias{calculate_rmse}
\title{Calculate rmse and mae.}
\usage{
calculate_rmse(all_forecasts, fits_long, T_rmse_min = 133, T_rmse_max = 243)
}
\arguments{
\item{all_forecasts}{tibble that contains row_number, model_id, variable, value and y_actual}

\item{fits_long}{tibble that contains pars_id, var_set and model_filename columns}

\item{T_rmse_min}{first T to calculate rmse}

\item{T_rmse_max}{last T to calculate rmse}
}
\value{
summary with rmse and mae for each model
}
\description{
Create summarising table with rmse and mae for each model.
}
\details{
Further checks are needed!
}
