% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{create_fits_long}
\alias{create_fits_long}
\title{Create fits_long data frame}
\usage{
create_fits_long(shifts, models, var_sets, horizons)
}
\arguments{
\item{shifts}{shifts data frame}

\item{models}{models data frame}

\item{var_sets}{var_sets data frame}

\item{horizons}{horizons data frame}
}
\value{
fits_long data frame
}
\description{
\code{create_fits_long} creates fits_long data frame.
}
\details{
Basically it is external product of \code{samples}, \code{var_sets} and \code{horizons} data frames plus 
creation of columns \code{result}, \code{model_filename}, \code{T_start_lower}.
}
\examples{
# TODO
}
