% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{create_interval_borders}
\alias{create_interval_borders}
\title{Create lower or upper interval border from future MCMC draws}
\usage{
create_interval_borders(y_future, level = c(80, 95), type = c("lower",
  "upper"))
}
\arguments{
\item{y_future}{data frame with n_draws rows. 
Each column represent future MCMC draws for specific variable for specific forecast horizon.}

\item{level}{vector of confidence levels}

\item{type}{lower or upper for left and right borders of intervals}
}
\value{
matrix of interval borders. One column corresponds for one confidence level.
}
\description{
\code{create_interval_borders} creates lower or upper interval borders from future MCMC draws
}
\details{
Just finds the appropriate quantile of future draws.
}
\examples{
a <- tibble::tibble(x = rnorm(100, mean = 2, sd = 1), 
  y = rnorm(100, mean = -2, sd = 4))
create_interval_borders(a, type = "lower")
}
