% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_ets}
\alias{estimate_ets}
\title{Estimate ETS model on multivariate time series}
\usage{
estimate_ets(y, ...)
}
\arguments{
\item{y}{multivariate time series}

\item{...}{further arguments passed to \code{ets} function}
}
\value{
list of estimated ETS models
}
\description{
\code{estimate_ets} returns a list of estimated ETS models
}
\details{
Just a wrapper for \code{auto.arima} function from \code{forecast} package.
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
ets_model <- estimate_ets(y_small)
}
