% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_rw}
\alias{estimate_rw}
\title{Dummy function to estimate random walk model}
\usage{
estimate_rw(y, ...)
}
\arguments{
\item{y}{multivariate time series}

\item{...}{dummy for further arguments}
}
\value{
message
}
\description{
\code{estimate_rw} estimates nothing at all :)
}
\details{
This function just provides common syntax interface and should be never used.
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
junk <- estimate_rw(y_small)
}
