% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_tvp_primiceri}
\alias{estimate_tvp_primiceri}
\title{Estimate TVP model on multivariate time series}
\usage{
estimate_tvp_primiceri(y, p = 12, nrep = 1000, nburn = 1000, ...)
}
\arguments{
\item{y}{multivariate time series}

\item{p}{number of lags}

\item{nrep}{number of MCMC draws excluding burn-in}

\item{nburn}{number of burn-in MCMC draws}

\item{...}{furhter arguments passed to \code{bvar.sv.tvp}}
}
\value{
estimated TVP model
}
\description{
\code{estimate_tvp_primiceri} returns an estimated TVP model a la Primiceri
}
\details{
Just a wrapper for \code{bvar.sv.tvp} function from \code{bvarsv} package.
ACHTUNG: probably we need to scale series for this model!
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
y_small_scaled <- scale_to(y_small)
primiceri_model <- estimate_tvp_primiceri(y_small_scaled, p = 1)
}
