% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_var}
\alias{estimate_var}
\title{Estimate VAR model on multivariate time series}
\usage{
estimate_var(y, p = 1, ...)
}
\arguments{
\item{y}{multivariate time series}

\item{p}{number of lags for VAR model}

\item{...}{further arguments passed to \code{VAR} function}
}
\value{
estimated VAR model
}
\description{
\code{estimate_var} returns estimated VAR model
}
\details{
Simply a wrapper for \code{VAR} function from \code{vars} package.
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
var_model <- estimate_var(y_small, p = 1)
}
