% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_var_lasso}
\alias{estimate_var_lasso}
\title{Estimate VAR-lasso model on multivariate time series}
\usage{
estimate_var_lasso(y, h = 1, p = 12, struct = "OwnOther", gran = c(25,
  10), T1 = floor(nrow(as.matrix(y))/3), T2 = floor(2 *
  nrow(as.matrix(y))/3))
}
\arguments{
\item{y}{multivariate time series}

\item{h}{forecasting horizon}

\item{p}{number of lags}

\item{struct}{type of VAR-lasso}

\item{gran}{granularity vector, If gran = (a, b) then cross-validation checks b values form lambda to lambda/a.}

\item{T1}{left border for cross validation samples}

\item{T2}{right border for cross validation samples}
}
\value{
estimated VAR-lasso model
}
\description{
\code{estimate_var_lasso} returns an estimated VAR-lasso model
}
\details{
Just a wrapper for \code{cv.BigVAR} function from \code{BigVAR} package. This function does NOT normalise variables. 
Pre-normalisation is required by the idea.
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
y_small_scaled <- scale_to(y_small)
var_lasso_model <- estimate_var_lasso(y_small_scaled)
}
