% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{forecast_all}
\alias{forecast_all}
\title{Forecast all requested models}
\usage{
forecast_all(fits_long, var_sets, ncores = 1, basefolder,
  log_file = "estim_log.txt", ...)
}
\arguments{
\item{fits_long}{data frame with requested models}

\item{var_sets}{correspondance between variable sets and variable names}

\item{ncores}{number of cores used. Just ignored.}

\item{basefolder}{path to folder where results are stored}

\item{log_file}{the name of log file}

\item{...}{further arguments passet to \code{forecast_one_fit}}
}
\value{
fits_long with estimation messages for each requested model.
}
\description{
\code{forecast_all} forecasts all requested models.
}
\details{
Just do it.
}
\examples{
# TODO
}
