% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_arima}
\alias{forecast_arima}
\title{Forecast using ARIMA model}
\usage{
forecast_arima(y, h = 1, model = NULL, ...)
}
\arguments{
\item{y}{multivariate time series}

\item{h}{forecast horizon}

\item{model}{list of estimated ARIMA models. If missing will be estimated automatically.}

\item{...}{further arguments passed to \code{auto.arima} function}
}
\value{
forecasts from ARIMA model as mforecast object
}
\description{
\code{forecast_arima} forecasts multivariate time series using ARIMA model
}
\details{
\code{auto.arima} function from \code{forecast} package is used
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
arima_forecast <- forecast_arima(y_small, h = 2)
}
