% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_ets}
\alias{forecast_ets}
\title{Forecast using ETS model}
\usage{
forecast_ets(y, h = 1, model = NULL, ...)
}
\arguments{
\item{y}{multivariate time series}

\item{h}{forecast horizon}

\item{model}{list of estimated ETS models. If missing will be estimated automatically.}

\item{...}{further arguments passed to \code{ets} function}
}
\value{
forecasts from ETS model as mforecast object
}
\description{
\code{forecast_ets} forecasts time series using ETS model
}
\details{
\code{ets} function from \code{forecast} package is used
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
ets_forecast <- forecast_ets(y_small, h = 2)
}
