% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{forecast_one_fit}
\alias{forecast_one_fit}
\title{Try to estimate and forecast particular request}
\usage{
forecast_one_fit(fit_row, var_sets, ts_data = torro::rus_macro,
  redundant = NULL)
}
\arguments{
\item{fit_row}{one row from data frame with models to estimate.}

\item{var_sets}{data frame with variable names. Should contain at least two columns:
\code{var_set} name of a set,
\code{variable} names of the variables in the set.}

\item{ts_data}{multivariate time series}

\item{redundant}{vector of column names NOT corresponding to parameters.
If NULL then built-in vector is used. See \code{get_parameter_colnames}.}
}
\value{
forecasts from estimated model or \code{try-error} class
}
\description{
\code{forecast_one_fit} takes one row of fits_long and tries to forecast with corresponding model.
To protect against possible errors with complex models \code{try} is used.
}
\details{
Does not use parallel computations.
}
\examples{
fits_long_toy <- create_fits_long(shifts_toy, models_toy, var_sets_toy, horizons_toy)
fit_row <- fits_long_toy[1, ]
attempt <- forecast_one_fit(fit_row, var_sets)
}
