% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_rw}
\alias{forecast_rw}
\title{Forecast using random walk model}
\usage{
forecast_rw(y, h = 1, ...)
}
\arguments{
\item{y}{multivariate time series}

\item{h}{forecast horizon}

\item{...}{further arguments passed to \code{rwf} function}
}
\value{
forecasts from RW model as mforecast object
}
\description{
\code{forecast_rw} forecasts time series using RW model
}
\details{
\code{rwf} function from \code{forecast} package is used
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
rw_forecast <- forecast_rw(y_small, h = 2)
}
