% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_tvp_primiceri}
\alias{forecast_tvp_primiceri}
\title{Forecast using TVP a la Primiceri model}
\usage{
forecast_tvp_primiceri(y, h = 1, model = NULL, level = c(80, 95), ...)
}
\arguments{
\item{y}{multivariate time series}

\item{h}{forecast horizon}

\item{model}{estimated TVP models. If missing will be estimated automatically.}

\item{level}{vector of confidence levels}

\item{...}{further arguments passed to \code{bvar.sv.tvp} function}
}
\value{
forecasts from TVP model as mforecast object
}
\description{
\code{forecast_tvp_primiceri} forecasts time series using TVP model
}
\details{
One dimensional forecasts are obtained from future MCMC draws. 
Mean and interval forecasts are calculated from the draws.
ACHTUNG: probably we need to scale series for this model!
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
y_small_scaled <- scale_to(y_small)
tvp_forecast <- forecast_tvp_primiceri(y_small_scaled, h = 2, p = 1)
}
