% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_var}
\alias{forecast_var}
\title{Forecast using VAR model}
\usage{
forecast_var(y, h = 1, model = NULL, ...)
}
\arguments{
\item{y}{multivariate time series}

\item{h}{forecast horizon}

\item{model}{estimated VAR model. If missing will be estimated automatically.}

\item{...}{further arguments passed to \code{estimate_var} and than to \code{VAR} function}
}
\value{
forecasts from VAR model as mforecast object
}
\description{
\code{forecast_var} forecasts time series using VAR model
}
\details{
Models are estimated using \code{VAR} function from \code{vars} package if necessary.
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
var_forecast <- forecast_var(y_small, h = 2, p = 1)
}
