% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_var_lasso}
\alias{forecast_var_lasso}
\title{Forecast using VAR-lasso model}
\usage{
forecast_var_lasso(y, h = 1, model = NULL, p = 1, struct = "OwnOther",
  gran = c(25, 10), type = c("fast", "honest"), h_cv = 1, ...)
}
\arguments{
\item{y}{multivariate time series}

\item{h}{forecast horizon}

\item{model}{estimated VAR-lasso model. If missing will be estimated automatically.}

\item{p}{number of lags}

\item{struct}{type of VAR-lasso}

\item{gran}{granularity vector, If gran = (a, b) then cross-validation checks b values form lambda to lambda/a.}

\item{type}{fast or honest cross-validation. 
honest: cross-validation is repeated for each h, so h should be a vector.
model should be missing or a list of the same length as h or the same model for all h.
fast: do cross-validation for h_cv and than predict for each horizon from 1 to h.}

\item{h_cv}{forecast horizon used for cross validation}

\item{...}{further arguments passed to \code{cv.BigVAR} function}
}
\value{
forecasts from VAR-lasso model as mforecast object
}
\description{
\code{forecast_var_lasso} forecasts time series using VAR-lasso model
}
\details{
\code{cv.BigVAR} function from \code{BigVAR} package is used
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
var_lasso_forecast <- forecast_var_lasso(y_small, h = 2)
}
