% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_evaluation.R
\name{get_forecasts_from_fit_files}
\alias{get_forecasts_from_fit_files}
\title{Read forecasts from fit files}
\usage{
get_forecasts_from_fit_files(fit_file_names = NULL, basefolder)
}
\arguments{
\item{fit_file_names}{the names of fit files. If not specified all files are used.}

\item{basefolder}{folder with subfolder fits with fit files.}
}
\value{
tibble with forecasts
}
\description{
\code{get_forecasts_from_fit_files} reads forecasts from fit files.
}
\details{
Read forecasts from fit files.
}
