% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{get_optimum_at_border_flag}
\alias{get_optimum_at_border_flag}
\title{Get optimum at border flag from fit file}
\usage{
get_optimum_at_border_flag(fit_file_name, basefolder)
}
\arguments{
\item{fit_file_name}{the name of fit file.}

\item{basefolder}{folder with subfolder fits with fit files.}
}
\value{
optimum at border flag, TRUE/FALSE.
}
\description{
\code{get_optimum_at_border_flag} gets optimum at border flag.
}
\details{
Recovers optimum at border flag from one fit file. Internal function.
}
