% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{get_optimum_at_border_flags}
\alias{get_optimum_at_border_flags}
\title{Get optimum at border flags}
\usage{
get_optimum_at_border_flags(fit_file_name, basefolder)
}
\arguments{
\item{fit_file_name}{fit file names.}

\item{basefolder}{folder with subfolder fits with fit files.}
}
\value{
vector of TRUE/FALSE.
}
\description{
\code{get_optimum_at_border_flags} gets optimum at border flags.
}
\details{
Recovers optimum at border flags from fit files.
}
\examples{
get_optimum_at_border_flags(c("fit_42.Rds", "fit_14.Rds"), tempdir())
}
