% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{get_scales}
\alias{get_scales}
\title{Means and standard deviations of time series}
\usage{
get_scales(y)
}
\arguments{
\item{y}{multivariate time series, matrix or data.frame}
}
\value{
tibble with three columns: variable, mean, sd
}
\description{
\code{get_scales} returns tibble with means and standard deviations of times series
}
\details{
The function calculates means and standard deviations of multivariate time series or data.frame object. 
Missing values are omitted. The result is output as tibble.
}
\examples{
get_scales(cars)
}
