% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{get_status_from_fit_file}
\alias{get_status_from_fit_file}
\title{Recover estimation status from fit file}
\usage{
get_status_from_fit_file(basefolder, fit_file_name)
}
\arguments{
\item{basefolder}{folder with subfolder fits with fit files.}

\item{fit_file_name}{the name of fit file.}
}
\value{
status message.
}
\description{
\code{get_status_from_fit_file} recovers estimation status from fit file.
}
\details{
Recovers estimation status from one fit file.
}
