% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{granularity_to_vector}
\alias{granularity_to_vector}
\title{Coerce gran_1 and gran_2 scalars in list into vector gran}
\usage{
granularity_to_vector(df)
}
\arguments{
\item{df}{list with gran_1 and gran_2 scalar numbers}
}
\value{
list with gran vector
}
\description{
\code{granulatiry_to_vector} coerces gran_1 and gran_2 scalars in list into vector gran in the same list.
}
\details{
Coerces gran_1 and gran_2 scalars in list into vector gran in the same list.
}
\examples{
a <- list(gran_1 = 25, gran_2 = 10)
granularity_to_vector(a)
}
