% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{data}
\name{horizons}
\alias{horizons}
\title{Horizons}
\format{tibble}
\usage{
data(horizons)
}
\description{
Tibble with the following column
}
\details{
\itemize{
\item h forecasting horizon
}
See also toy version, \code{horizons_toy}.
}
\keyword{datasets}
