% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{data}
\name{horizons_toy}
\alias{horizons_toy}
\title{Horizons toy version}
\format{tibble}
\usage{
data(horizons_toy)
}
\description{
Tibble with the following column
}
\details{
\itemize{
\item h forecasting horizon
}
See also full version, \code{horizons}.
}
\keyword{datasets}
