% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{is_optimum_at_border}
\alias{is_optimum_at_border}
\title{Check whether BigVAR LASSO optimum landed at border}
\usage{
is_optimum_at_border(fit)
}
\arguments{
\item{fit}{object of type mforecast.}
}
\value{
TRUE/FALSE.
}
\description{
\code{is_optimum_at_border} checks whether BigVAR LASSO optimum landed at border.
}
\details{
Takes mforecast objects, extracts model.
Than checks whether optimum is equal to the minimal lambda considered.
Returns FALSE for non BigVAR LASSO models.
}
\examples{
fits_long_toy <- create_fits_long(shifts_toy, models_toy, var_sets_toy, horizons_toy)
fit_row <- fits_long_toy[1, ]
attempt <- forecast_one_fit(fit_row, var_sets)
is_optimum_at_border(attempt)
}
