% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{log_message}
\alias{log_message}
\title{Log message for boring estimation}
\usage{
log_message(fit_row, full_path_to_log_file, additional_message = "")
}
\arguments{
\item{fit_row}{one row data.frame with at least model_type column}

\item{full_path_to_log_file}{maybe empty}

\item{additional_message}{additional message}
}
\value{
message message
}
\description{
Message is returned and (possibly) written to file.
}
\details{
May include additional message.
}
\examples{
fits_long_toy <- create_fits_long(shifts_toy, models_toy, var_sets_toy, horizons_toy)
log_message(fits_long_toy[1, ], tempfile(), "Hi, there!\\n")
}
