% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_misc.R
\name{matrix_to_mforecast}
\alias{matrix_to_mforecast}
\title{Transform simple matrix of forecasts to mforecast object.}
\usage{
matrix_to_mforecast(forecast_matrix, y_before, method = "Unspecified")
}
\arguments{
\item{forecast_matrix}{h by m forecast matrix}

\item{y_before}{tibble with three columns: variable, mean, sd}

\item{method}{the name of the method}
}
\value{
complex mforecast object
}
\description{
\code{matrix_to_mforecast} transform simple matrix of forecasts to complex mforecast object.
}
\details{
\code{forecast} package uses complex mforecast format to store forecasts.
This function transforms simple matrices to mforecast objects.
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
y_forecast <- tail(y_small, 1) # trivial forecast 
mfor <- matrix_to_mforecast(y_forecast, y_small, method = "Trivial")
# ggplot2:::autoplot.mforecast(mfor) # TODO: does not work!!!
}
