% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_misc.R
\name{mforecast_to_matrix}
\alias{mforecast_to_matrix}
\title{Transform mforecast object to simple matrix of forecasts.}
\usage{
mforecast_to_matrix(mforecast)
}
\arguments{
\item{mforecast}{mforecast object to be transformed}
}
\value{
h by m matrix (h — number of steps, m — number of variables)
}
\description{
\code{mforecast_to_matrix} transform complex mforecast object to simple matrix of forecasts.
}
\details{
\code{forecast} package uses complex mforecast format to store forecasts.
This function transforms mforecast objects to simple matrices.
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
var_model <- vars::VAR(y_small)
y_for <- forecast::forecast(var_model, h = 2)
mforecast_to_matrix(y_for)
}
