% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_misc.R
\name{next_obs_time}
\alias{next_obs_time}
\title{Get time of the next to last observation of time series}
\usage{
next_obs_time(y)
}
\arguments{
\item{y}{time series}
}
\value{
time of the last observation plus one time unit
}
\description{
\code{next_obs_time} gives the time of the next to last observation of time series
}
\details{
Let's consider monthly time series. In the output 2015.0 means 2015 first month, 
2015.083 means 2015 second month and so on.
}
\examples{
data(rus_macro)
next_obs_time(rus_macro)
}
