% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{primiceri_draws_to_1d_forecast}
\alias{primiceri_draws_to_1d_forecast}
\title{Convert future MCMC draws from TVP a la Primiceri to interval borders}
\usage{
primiceri_draws_to_1d_forecast(model, y_actual, series_no = 1, h = 1,
  level = c(80, 95))
}
\arguments{
\item{model}{estimated TVP a la primiceri model}

\item{y_actual}{original multivariate time series. One column is extracted and is stored in forecast object.}

\item{series_no}{number of time series to predict}

\item{h}{forecasting horizon}

\item{level}{vector of confidence levels}
}
\value{
one dimensional complex forecast object
}
\description{
\code{primiceri_draws_to_1d_forecast} creates lower and upper interval borders, mean  
from future MCMC draws a la Primiceri for one specific variable.
}
\details{
For every h future MCMC draws are obtained, than standard forecast object is created.
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
y_small_scaled <- scale_to(y_small)
primiceri_model <- estimate_tvp_primiceri(y_small_scaled, p = 1)
draws <- primiceri_draws_to_1d_forecast(primiceri_model, 
  y_actual = y_small_scaled, h = 1, series_no = 1)
}
