% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{reforecast_fit}
\alias{reforecast_fit}
\title{Reforecast mforecast object on new horizon}
\usage{
reforecast_fit(old_fit, new_h = 1)
}
\arguments{
\item{old_fit}{mforecast object}

\item{new_h}{new forecast horizon}
}
\value{
forecasts as mforecast object
}
\description{
\code{reforecast_fit} reforecast mforecast object on new horizon.
}
\details{
Currently works for auto.arima, ets, fast version of bigvar_lasso.
}
\examples{
data(rus_macro)
y_small <- rus_macro[, c("cpi", "employment", "m2")]
var_lasso_forecast <- forecast_var_lasso(y_small, h = 2)
var_lasso_reforecast <- reforecast_fit(var_lasso_forecast, new_h = 4)
}
