% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{data}
\name{rus_macro}
\alias{rus_macro}
\title{Russian macro indicators}
\format{Multivariate time-series, monthly, from 1995 to 2015, 23 variables}
\usage{
data(rus_macro)
}
\description{
Multivariate time-series with 23 variables. Monthly, deseasonalised. From 1995 to 2015.
}
\details{
\itemize{
\item construction 
\item cpi
\item employment
\item export
\item gas_price
\item gov_balance
\item ib_rate
\item import
\item ind_prod
\item labor_request
\item lend_rate
\item m2
\item ner
\item nfa_cb
\item oil_price
\item ppi
\item real_income
\item real_investment
\item reer
\item retail
\item unemp_rate
\item wage
\item agriculture
}
}
\keyword{datasets}
