% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_to}
\alias{scale_to}
\title{Scale multivariate time series to specific means and standard deviations}
\usage{
scale_to(y, mu_sd = NULL, dirty_scale = FALSE)
}
\arguments{
\item{y}{multivariate time series to be scales}

\item{mu_sd}{tibble with three columns: variable, mean, sd}

\item{dirty_scale}{logical, if TRUE then the series is assumed having mean = 0 and sd = 1,
and is just multiplied by sd and augmented by mean.}
}
\value{
scaled mutlivatiate time series
}
\description{
\code{scale_to} scales multivariate time series to specific means and standard deviations
}
\details{
The function scales multivariate time series to specific means and standard deviations.
If mu_sd tibble is not specified then times series are scaled to zero mean and unit standard deviation.
}
\examples{
scale_to(cars)
data(rus_macro)
scale_to(rus_macro)
}
