% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{set_agnostic_max_h}
\alias{set_agnostic_max_h}
\title{Set maximal horizon for h-agnostic models}
\usage{
set_agnostic_max_h(fits_long, ignored_vars = "model_filename")
}
\arguments{
\item{fits_long}{list of models to estimate}

\item{ignored_vars}{names of variables that are ignored 
when \code{set_agnostic_max_h} groups rows correponding to same model.}
}
\value{
shorter list of models to estimate
}
\description{
\code{set_agnostic_max_h} sets maximal horizon for h-agnostic models.
}
\details{
There are two types of models. Horizon agnostic that do not require horizon h for estimation (ARIMA, RW, ETS...). 
Horizon gnostic models that require h both for estimation and forecasting (cross validated VAR-lasso, ...). 
This function removes redundant estimation of h-agnostic models. If one requires estimation of h-agnostic
model for h = 1, h = 2, h = 3 this function will leave only estimation of h = 3 in the request.
List of models to estimate should contain at least
\itemize{
\item h forecasting horizon
\item h_required logical TRUE/FALSE, FALSE for h-agnostic models
}
All other variables except \code{ignored_vars} are used to identify whether models are different.
}
\examples{
fits_long <- tibble::tibble(h_required = c(TRUE, TRUE, FALSE, FALSE, FALSE),
   h = c(1, 2, 1, 2, 3), params = c(1, 1, 2, 2, 5))
set_agnostic_max_h(fits_long)
}
