% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{data}
\name{shifts_toy}
\alias{shifts_toy}
\title{Shifts toy version}
\format{tibble}
\usage{
data(shifts_toy)
}
\description{
Tibble with the following columns
}
\details{
\itemize{
\item shift_name name of the shift
\item shift_T_start starting time of the shift
\item win_expanding logical, TRUE for expanding/growing window, FALSE for moving
\item win_start_lenght length of the first window
\item n_shifts number of window shifts
}
See also full version, \code{shifts}.
}
\keyword{datasets}
