% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{package}
\name{torro}
\alias{torro}
\alias{torro-package}
\title{torro: project-package to forecast russian macro indicators with BVARs and alternative models}
\description{
Trying to test project-package idea.
}
\details{
Trying to test project-package idea...
}
