% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_evaluation.R
\name{transform_to_relative_measure}
\alias{transform_to_relative_measure}
\title{Transform absolute error measures to relative.}
\usage{
transform_to_relative_measure(rmse_table, the_model_type = "rw",
  the_pars_id = "automatic")
}
\arguments{
\item{rmse_table}{output of `calculate_rmse` function.
Should contain columns: row_number /horizon/, variable, var_set, model_type, pars_id
and some measures like mse, rmse etc}

\item{the_model_type}{base model type}

\item{the_pars_id}{base parameter set}
}
\value{
measure table with relative measures
}
\description{
Transform absolute error measures to relative.
}
\details{
Transform absolute error measures to relative.
}
