% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torro.R
\docType{data}
\name{var_sets}
\alias{var_sets}
\title{Variable sets}
\format{tibble}
\usage{
data(var_sets)
}
\description{
Tibble with the following columns
}
\details{
\itemize{
\item var_set the name of variable set
\item variable names of variables in the variable set
\item pre_transform transformation before estimation (TODO)
\item post_transform transformation of forecasts (TODO)
}
See also toy version, \code{var_sets_toy}.
}
\keyword{datasets}
