% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{attempt_to_status}
\alias{attempt_to_status}
\title{Extract message from forecasting attempt}
\usage{
attempt_to_status(attempt)
}
\arguments{
\item{attempt}{estimation attempt.}
}
\value{
character message
}
\description{
\code{attempt_to_status} extracts error message from estimation attempt.
}
\details{
Gives "OK" if attempt was successful.
}
\examples{
fits_long_toy <- create_fits_long(shifts_toy, models_toy, var_sets_toy, horizons_toy)
fit_row <- fits_long_toy[1, ]
attempt <- forecast_one_fit(fit_row, var_sets)
attempt_to_status(attempt)
}
