% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_misc.R
\name{get_status_from_fit_files}
\alias{get_status_from_fit_files}
\title{Recover estimation status from fit files}
\usage{
get_status_from_fit_files(fit_file_name, basefolder)
}
\arguments{
\item{fit_file_name}{fit file names.}

\item{basefolder}{folder with subfolder fits with fit files.}
}
\value{
status messages.
}
\description{
\code{get_status_from_fit_files} recovers estimation status from fit file.
}
\details{
Recovers estimation status from fit files.
Vectorised version of \code{get_status_from_fit_file()}.
}
\examples{
get_status_from_fit_files(c("fit_42.Rds", "fit_14.Rds"), tempdir())
}
